bg(db): Blocking unauthorized access
You can't review a session which doesn't exist
You can't review the same session twice
Admin only can change or delete
Should request session to only mentors and who exists
The same question shouldn't be send to the same Mentor.
you should be a mentor to accept/reject session and it should be addressed to you!

[##168426458]